ai-server/pyproject.toml
[project]
name = "ai-server"
version = "0.1.0"
description = "Server for NAILA Robot Prototype"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.11.0",
    "factory-boy>=3.3.3",
    "faster-whisper>=1.2.1",
    "freezegun>=1.5.5",
    "langchain>=1.0.3",
    "langgraph>=1.0.2",
    "llama-cpp-python>=0.3.16",
    "num2words>=0.5.14",
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "paho-mqtt>=2.1.0",
    "pillow>=12.0.0",
    "piper-tts>=1.3.0",
    "psutil>=7.1.3",
    "py-cpuinfo>=9.0.0",
    "pydub>=0.25.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.8.0",
    "resampy>=0.4.3",
    "sentence-transformers>=5.1.2",
    "soundfile>=0.13.1",
    "structlog>=25.5.0",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "ultralytics>=8.3.225",
]
